components/hf_endpoint_component/requirements.txt
biopython==1.83
pyarrow==15.0.0
python-dotenv==1.0.1
requests==2.31.0
google-cloud-storage==2.15.0